+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -339,6 +339,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed a null reference exception in the probe UI when no HDRP asset is present.
 - Fixed the outline example in the doc (sampling range was dependent on screen resolution)
 - Fixed a null reference exception in the HDRI Sky editor when no HDRP asset is present.
+- Fixed Manual Depth of field on Metal platforms.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled